+++ src/main/java/edu/illinois/library/cantaloupe/image/Metadata.java
@@ -8,6 +8,7 @@ import edu.illinois.library.cantaloupe.Application;
 import edu.illinois.library.cantaloupe.image.exif.Directory;
 import edu.illinois.library.cantaloupe.image.exif.Tag;
 import edu.illinois.library.cantaloupe.image.iptc.DataSet;
+import edu.illinois.library.cantaloupe.image.xmp.MapReader;
 import edu.illinois.library.cantaloupe.util.StringUtils;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
@@ -17,6 +18,7 @@ import org.apache.jena.riot.RiotException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.io.StringReader;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
@@ -170,14 +172,30 @@ public class Metadata {
     }
 
     /**
-     * Returns an RDF/XML string in UTF-8 encoding. The root element is
-     * {@literal rdf:RDF}, and there is no packet wrapper.
-     *
-     * @return XMP data packet.
+     * @return RDF/XML string in UTF-8 encoding. The root element is {@literal
+     *         rdf:RDF}, and there is no packet wrapper.
      */
     @JsonProperty
     public Optional<String> getXMP() {
         return Optional.ofNullable(xmp);
+    }
+
+    /**
+     * @return Map of elements found in the XMP data. If none are found, the
+     *         map is empty.
+     */
+    @JsonIgnore
+    public Map<String,Object> getXMPElements() {
+        loadXMP();
+        if (xmpModel != null) {
+            try {
+                MapReader reader = new MapReader(xmpModel);
+                return reader.readElements();
+            } catch (IOException e) {
+                LOGGER.warn("getXMPElements(): {}", e.getMessage());
+            }
+        }
+        return Collections.emptyMap();
     }
 
     /**
@@ -290,7 +308,9 @@ public class Metadata {
      * {
      *     "exif": See {@link Directory#toMap()},
      *     "iptc": See {@link DataSet#toMap()},
-     *     "xmp": "<rdf:RDF>...</rdf:RDF>",
+     *     "xmp_string": "<rdf:RDF>...</rdf:RDF>",
+     *     "xmp_model": [Jena model],
+     *     "xmp_elements": {@link Map}
      *     "native": String
      * }}
      *
@@ -314,6 +334,7 @@ public class Metadata {
         // XMP
         getXMP().ifPresent(xmp -> map.put("xmp_string", xmp));
         getXMPModel().ifPresent(model -> map.put("xmp_model", model));
+        map.put("xmp_elements", getXMPElements());
         // Native metadata
         getNativeMetadata().ifPresent(nm -> map.put("native", nm));
         return Collections.unmodifiableMap(map);