+++ src/main/java/edu/illinois/library/cantaloupe/image/xmp/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * <p>Contains an XMP parser.</p>
+ *
+ * @since 6.0
+ */
+package edu.illinois.library.cantaloupe.image.xmp;